Added modules for campgrounds and modified readme
