+++ README.md
@@ -3,11 +3,3 @@
 * this repo is based upon a project from Colt Steele's Web Developer Bootcamp
 * I am really thankful to Colt, I learnt so much
 * Keep doing the good work brother!
-
-# Routes
-   name      url           verb           description
-======================================================
-* INDEX     /dogs           GET     Display a list of all dogs
-* NEW       /dogs/new       GET     Displays a form to make a new dog
-* CREATE    /dogs           POST    Add new dog to DB
-* SHOW      /dogs/:id       GET     Show info about one dog
