+++ app.js
@@ -1,20 +1,13 @@
 var express     = require("express"),
     app         = express(),
     bodyParser  = require('body-parser'),
-    mongoose    = require('mongoose');
+    mongoose    = require('mongoose'),
+    Campground   = require('./models/campground');
 
 mongoose.connect("mongodb://localhost:27017/yelp_camp",{useNewUrlParser: true, useUnifiedTopology: true});
 app.use(bodyParser.urlencoded({extended: true}));
 app.set("view engine","ejs");
-
-// Schema Setup
-var campgroundSchema = new mongoose.Schema({
-    name: String,
-    image: String, 
-    description: String
-});
-
-var Campground = mongoose.model("Campground", campgroundSchema);        
+    
 // Campground <==> db.campgrounds
 
 // Campground.create(
