+++ models/campground.js
@@ -1,0 +1,11 @@
+var mongoose = require('mongoose');
+
+// Schema Setup
+var campgroundSchema = new mongoose.Schema({
+    name: String,
+    image: String, 
+    description: String
+});
+
+var Campground = mongoose.model("Campground", campgroundSchema);    
+module.exports = Campground;
